% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{gpt3.make_request}
\alias{gpt3.make_request}
\title{Make a test request to the GPT-3 API}
\usage{
gpt3.make_request(
  prompt_,
  model_ = "text-davinci-002",
  output_type_ = "string_only",
  suffix_ = NULL,
  max_tokens_ = 256,
  temperature_ = 0.9,
  top_p_ = 1,
  n_ = 1,
  stream_ = F,
  logprobs_ = NULL,
  echo_ = F,
  stop_ = NULL,
  presence_penalty_ = 0,
  frequency_penalty_ = 0,
  best_of_ = 1,
  logit_bias_ = NULL
)
}
\arguments{
\item{verbose}{(boolean) if TRUE prints the actual prompt and GPT-3 completion of the test request (default: FALSE).}
}
\value{
A message of success or failure of the connection.
}
\description{
\code{gpt3.test_request()} sends a basic \href{https://beta.openai.com/docs/api-reference/completions}{completion request} to the Open AI GPT-3 API.
}
\examples{
gpt3.test_request()
}
