% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_request.R
\name{gpt3.test_request}
\alias{gpt3.test_request}
\title{Make a test request to the GPT-3 API}
\usage{
gpt3.test_request(verbose = F)
}
\arguments{
\item{verbose}{(boolean) if TRUE prints the actual prompt and GPT-3 completion of the test request (default: FALSE).}
}
\value{
A message of success or failure of the connection.
}
\description{
\code{gpt3.test_request()} sends a basic \href{https://beta.openai.com/docs/api-reference/completions}{completion request} to the Open AI GPT-3 API.
}
\examples{
gpt3.test_request()
}
