% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_completion.R
\name{gpt3_test_completion}
\alias{gpt3_test_completion}
\title{Make a test request to the GPT-3 API}
\usage{
gpt3_test_completion(verbose = T)
}
\arguments{
\item{verbose}{(boolean) if TRUE prints the actual prompt and GPT-3
completion of the test request (default: TRUE).}
}
\value{
A message of success or failure of the connection.
}
\description{
\code{gpt3_test_completion()} sends a basic \href{https://beta.openai.com/docs/api-reference/completions}{completion request} to the
Open AI GPT-3 API.
}
\examples{
gpt3_test_completion()
}
