% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpt_embeddings.R
\name{rgpt_embeddings}
\alias{rgpt_embeddings}
\title{Retrieves text embeddings for character input from a vector from OpenAI's GPT API}
\usage{
rgpt_embeddings(
  input_var,
  id_var,
  param_model = "text-embedding-3-large",
  param_dimensions = 256
)
}
\arguments{
\item{input_var}{character vector that contains the texts for which you want to obtain text embeddings from the specified GPT model}

\item{id_var}{(optional) character vector that contains the user-defined ids of the prompts. See details.}

\item{param_model}{a character vector that indicates the \href{https://beta.openai.com/docs/guides/embeddings/embedding-models}{embedding model}; one of "text-embedding-3-large" (default), "text-embedding-3-small", "text-embedding-ada-002"}

\item{param_dimensions}{a numeric value (default: 256) to indicate the number of dimensions to shorten the embeddings size (see: https://platform.openai.com/docs/guides/embeddings). This parameter only works for the \code{text-embedding-3-large} model. For all embeddings dimensions, set the parameter to 3072.}
}
\value{
A data.table with the embeddings as separate columns; one row represents one input text. See details.
}
\description{
\code{rgpt_embeddings()} extends the single embeddings function \code{rgpt_single_embedding()} to allow for the processing of a whole vector
}
\details{
The returned data.table contains the column \code{id} which indicates the text id (or its generic alternative if not specified) and the columns \code{dim_1} ... \verb{dim_\{max\}}, where \code{max} is the length of the text embeddings vector that the different models (see below) return. For the default "Embedding V3 large" model, these are 3072 dimensions (i.e., \code{dim_1}... \code{dim_3072}).

The function supports the text similarity embeddings for the \href{https://platform.openai.com/docs/models/embeddings}{three GPT embeddings models} as specified in the parameter list.
\itemize{
\item Embedding V3 large \code{text-embedding-3-large} (3072 dimensions)
\item Embedding V3 small \code{text-embedding-3-small} (1536 dimensions)
\item Ada 2nd generation \code{text-embedding-ada-002} (1536 dimensions)
}

Note that the dimension size (= vector length), speed and \href{https://openai.com/api/pricing/}{associated costs} differ considerably.

These vectors can be used for downstream tasks such as (vector) similarity calculations.
}
\examples{
# First authenticate with your API key via `rgpt_authenticate('pathtokey')`

# Use example data:
## The data below were generated with the `rgpt_single()` function as follows:
##### DO NOT RUN #####
# travel_blog_data = rgpt_single(prompt_content = "Write a travel blog about a dog's journey through the UK:", temperature = 0.8, n = 10, max_tokens = 200, seed = 36)[[1]]
##### END DO NOT RUN #####

# You can load these data with:
data("travel_blog_data") # the dataset contains 10 completions for the above request

## Obtain text embeddings for the completion texts:
emb_travelblogs = rgpt_embeddings(input_var = travel_blog_data$gpt_content)
dim(emb_travelblogs)
}
