% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpt_single_embedding.R
\name{rgpt_single_embedding}
\alias{rgpt_single_embedding}
\title{Obtains text embeddings for a single character (string) from OpenAI's GPT API}
\usage{
rgpt_single_embedding(input, model = "text-embedding-3-large")
}
\arguments{
\item{input}{character that contains the text for which you want to obtain text embeddings from the specified GPT model}

\item{model}{a character vector that indicates the \href{https://beta.openai.com/docs/guides/embeddings/embedding-models}{embedding model}; one of "text-embedding-3-large" (default), "text-embedding-3-small", "text-embedding-ada-002"}
}
\value{
A numeric vector (= the embedding vector)
}
\description{
\code{rgpt_single_embedding()} sends a single \href{https://beta.openai.com/docs/guides/embeddings}{embedding request} to the Open AI GPT API.
}
\details{
The returned data.table contains the column \code{id} which indicates the text id (or its generic alternative if not specified) and the columns \code{dim_1} ... \verb{dim_\{max\}}, where \code{max} is the length of the text embeddings vector that the different models (see below) return. For the default "Embedding V3 large" model, these are 3072 dimensions (i.e., \code{dim_1}... \code{dim_3072}).

The function supports the text similarity embeddings for the \href{https://platform.openai.com/docs/models/embeddings}{three GPT embeddings models} as specified in the parameter list.
\itemize{
\item Embedding V3 large \code{text-embedding-3-large} (3072 dimensions)
\item Embedding V3 small \code{text-embedding-3-small} (1536 dimensions)
\item Ada 2nd generation \code{text-embedding-ada-002} (1536 dimensions)
}

Note that the dimension size (= vector length), speed and \href{https://openai.com/api/pricing/}{associated costs} differ considerably.

These vectors can be used for downstream tasks such as (vector) similarity calculations.
}
\examples{
# First authenticate with your API key via `rgpt_authenticate('pathtokey')`

# Once authenticated:

## Simple request with defaults:
sample_string = "London is one of the most liveable cities in the world. The city is always full of energy and people. It's always a great place to explore and have fun."
rgpt_single_embedding(input = sample_string)

## Change the model:
rgpt_single_embedding(input = sample_string
  , model = 'text-embedding-ada-002')
}
