% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_numeric}
\alias{to_numeric}
\title{Convert character vector of numeric values into a numeric vector}
\usage{
to_numeric(x)
}
\arguments{
\item{x}{a character vector of numeric values}
}
\value{
A numeric vector
}
\description{
Converts a character vector of numeric values into a numeric vector
}
\examples{
to_numeric('12312')
}
