% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{travel_blog_data}
\alias{travel_blog_data}
\title{Travel Blog Data (template)}
\format{
\subsection{\code{travel_blog_data}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
travel_blog_data
}
\description{
Here give some further info ...
}
\keyword{datasets}
