% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_urls.R
\docType{data}
\name{url.completions}
\alias{url.completions}
\title{Contains the package's base URLs}
\format{
An object of class \code{character} of length 1.
}
\usage{
url.completions
}
\description{
These are the base URLs for the \code{rgpt3} package. Do not change these!
}
\keyword{datasets}
